added MJD functions, removed missed long doubles

+++ Kit/Include/orbkit.h
@@ -244,7 +244,7 @@ void PlanetEphemerides(long i, double JD, double mu, double *SMA, double *ecc,
                        double *anom, double *p, double *alpha, double *rmin,
                        double *MeanMotion, double *Period);
 void LunaPosition(double JD, double r[3]);
-void LunaInertialFrame(long double JulDay, double CNJ[3][3]);
+void LunaInertialFrame(double JulDay, double CNJ[3][3]);
 double LunaPriMerAng(double JulDay);
 void FindCLN(double r[3], double v[3], double CLN[3][3], double wln[3]);
 void FindCEN(double r[3], double CEN[3][3]);
